psyche/backend/requirements.txt
fastapi==0.109.2
uvicorn==0.27.1
pillow==11.0.0
python-multipart==0.0.9
watchdog==3.0.0
websockets==12.0 